agents/knowledge_rag_agent/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alibabacloud-credentials==1.0.1
alibabacloud-credentials-api==1.0.0
alibabacloud-darabonba-string==0.0.4
alibabacloud-kms20160120==2.2.3
alibabacloud-tea==0.4.3
alibabacloud_darabonba_array==0.1.0
alibabacloud_darabonba_encode_util==0.0.2
alibabacloud_darabonba_map==0.0.1
alibabacloud_darabonba_signature_util==0.0.4
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_pop==0.0.8
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.14
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
build==1.2.2.post1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
cryptography==44.0.2
fastapi==0.115.12
fastembed==0.6.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
grpcio==1.71.0
h11==0.16.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
loguru==0.7.3
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.3
nacos-client-python @ file:///Users/jin/DEVELOP/AI/libs/nacos-client-python/dist/nacos_client_python-0.1.0-py3-none-any.whl#sha256=cddd51eca066e7964845e5a5d29889ce75d6fefa156065ac20febb990aebc66f
nacos-sdk-python==2.0.4
numpy==2.2.5
onnxruntime==1.21.1
packaging==25.0
pillow==11.2.1
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
py_rust_stemmers==0.1.5
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyproject_hooks==1.2.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
